Unix-like systems compilation (debug and release, with all tests passing and benchmarks running).
Notes:
- Available only for gcc (tested with v12.2.0). Clang compiler (tested with v15.0.0) is still missing some C++20 features used in this project.
- Address sanitizer is disabled for the debug builds.

Applied suggestions coming from CLion's Inspect Code option:
- Added [[maybe_unused]].
- Added [[nodiscard]].
- Added explicit for constructors capable of taking only one argument.
- Added override for destructors of derived classes.
- Changed methods to static.
- Moved fmt formatter definitions prior to their first use.
- Removed constexpr from some function signatures.
- Removed virtual and added override for methods of derived classes.
- Removed some include headers.

Other minor changes:
- Added date package.
- Added SQLite3 package (containing sqlite3.h and sqlite3.c).
- Changed <fmt>, <gmock> and <gtest> header includes to "fmt", "gmock" and "gtest" respectively.
- Updated fmt package (version 9.1.0 contains some more chrono types formatters, such as fmt::weekday).
- Updated rtc package (with the unixlike-builds branch version).
- Updated ZipLib package.
- Updated perceptualdiff package.

+++ src/chapter_03_strings_and_regular_expressions/problem_031_transforming_dates_in_strings.cpp
@@ -1,10 +1,10 @@
 #include "chapter_03_strings_and_regular_expressions/problem_031_transforming_dates_in_strings.h"
 
+#include "fmt/ostream.h"
+
 #include <exception>
-#include <fmt/ostream.h>
 #include <iostream>  // cout
 #include <regex>  // regex_match, regex_replace, smatch
-#include <sstream>  // ostringstream
 #include <string>  // stoi
 #include <string_view>
 
@@ -22,9 +22,9 @@ namespace tmcppc::problem_31 {
             std::regex_match(s, matches, valid_format_pattern_2)) and
             matches.size() == 4) {
 
-            day = std::stoi(matches[1].str());
-            month = std::stoi(matches[2].str());
-            year = std::stoi(matches[3].str());
+            day = static_cast<uint16_t>(std::stoi(matches[1].str()));
+            month = static_cast<uint16_t>(std::stoi(matches[2].str()));
+            year = static_cast<uint16_t>(std::stoi(matches[3].str()));
         } else {
             throw invalid_date_format_error(s);
         }
